refactor: common 패키지로 BaseEntity, Exception 클래스 이동
