+++ src/main/java/yapp/buddycon/common/domain/BaseEntity.java
@@ -25,4 +25,5 @@ public class BaseEntity {
   @LastModifiedDate
   @Column(name = "updated_at")
   private LocalDateTime updatedAt = LocalDateTime.now();
+
 }
