+++ src/main/java/yapp/buddycon/common/exception/CustomException.java
@@ -7,4 +7,5 @@ import lombok.Getter;
 @AllArgsConstructor
 public class CustomException extends RuntimeException {
   private final ErrorCode errorCode;
+
 }
